Move validate_dataset_exists function to native scripts utils
