+++ components/clp-package-utils/clp_package_utils/scripts/native/search.py
@@ -22,11 +22,11 @@ from clp_package_utils.general import (
     CLP_DEFAULT_CONFIG_FILE_RELATIVE_PATH,
     get_clp_home,
     load_config_file,
-    validate_dataset_exists,
 )
 from clp_package_utils.scripts.native.utils import (
     run_function_in_process,
     submit_query_job,
+    validate_dataset_exists,
     wait_for_query_job,
 )
